Ya esta el archivo en su lugar correcto  - Arregle algunas cosas que fallaron en el merge.
