+++ src/main/test/edu/fiuba/algo3/entrega_1/Entrega1.java
@@ -1,36 +1,440 @@
-
 package edu.fiuba.algo3.entrega_1;
 
 import Entidades.AlgoRoma;
 import Entidades.Elementos.Comida;
-import Entidades.Elementos.DispositivoDeAzar;
 import Entidades.Elementos.Interactuable;
 import Entidades.Elementos.MockDado;
-import Entidades.Equipo.*;
-import Entidades.Errores.*;
+import Entidades.Energia.Energia;
+import Entidades.Errores.CantidadMinimaDeJugadores;
+import Entidades.Errores.PartidaFinalizada;
+import Entidades.Errores.SinDispositivoDeAzar;
 import Entidades.Jugadores.Gladiador;
 import Entidades.Jugadores.Jugador;
-import Entidades.ListaCircular;
-import Entidades.Obstaculos.Bacanal;
 import Entidades.Obstaculos.FieraSalvaje;
-import Entidades.Obstaculos.Piedra;
 import Entidades.Tablero.*;
-import Entidades.Energia.Energia;
+import Entidades.Equipo.*;
 import org.junit.Test;
 
 import static org.junit.jupiter.api.Assertions.*;
 
-public class Entrega1 {
+public class Entrega1{
+    public Mapa mapaConFieraSalvaje() {
+        int cantidadCasilleros = 30;
+        Interactuable[][] elementosMapa = new Interactuable[cantidadCasilleros][1];
 
+        for (int i = 0; i < cantidadCasilleros; i++) {
+            elementosMapa[i][0] = new FieraSalvaje();
+        }
+
+        InformacionMapaEnMatriz informacionMapaEnMatriz = new InformacionMapaEnMatriz(elementosMapa);
+
+        return new MapaLineal(informacionMapaEnMatriz);
+    }
+
+    public Mapa mapaConComida() {
+        int cantidadCasilleros = 10;
+        Interactuable[][] elementosMapa = new Interactuable[cantidadCasilleros][1];
+
+        for (int i = 0; i < cantidadCasilleros; i++) {
+            elementosMapa[i][0] = new Comida();
+        }
+
+        InformacionMapaEnMatriz informacionMapaEnMatriz = new InformacionMapaEnMatriz(elementosMapa);
+
+        return new MapaLineal(informacionMapaEnMatriz);
+    }
+
+    public Mapa mapaConCasco() {
+        int cantidadCasilleros = 10;
+        Interactuable[][] elementosMapa = new Interactuable[cantidadCasilleros][1];
+
+        for (int i = 0; i < cantidadCasilleros; i++) {
+            elementosMapa[i][0] = new Casco();
+        }
+
+        InformacionMapaEnMatriz informacionMapaEnMatriz = new InformacionMapaEnMatriz(elementosMapa);
+
+        return new MapaLineal(informacionMapaEnMatriz);
+    }
+
+    public Mapa mapa4CasillerosConCascoYFieraSalvaje() {
+        int cantidadCasilleros = 4;
+        Interactuable[][] elementosMapa = new Interactuable[cantidadCasilleros][1];
+
+        elementosMapa[0][0] = new Casco();
+        elementosMapa[1][0] = new Casco();
+        elementosMapa[2][0] = new FieraSalvaje();
+        elementosMapa[3][0] = new Casco();
+
+        InformacionMapaEnMatriz informacionMapaEnMatriz = new InformacionMapaEnMatriz(elementosMapa);
+
+        return new MapaLineal(informacionMapaEnMatriz);
+    }
+
+    public Mapa mapaConEquipamiento() {
+        int cantidadCasilleros = 10;
+        Interactuable[][] elementosMapa = new Interactuable[cantidadCasilleros][1];
+
+        elementosMapa[0][0] = new Casco();
+        elementosMapa[1][0] = new Casco();
+        elementosMapa[2][0] = new Armadura();
+        elementosMapa[3][0] = new EscudoYEspada();
+        elementosMapa[4][0] = new Llave();
+        elementosMapa[5][0] = new FieraSalvaje();
+        elementosMapa[6][0] = new Casco();
+
+        for (int i = 7; i < cantidadCasilleros; i++) {
+            elementosMapa[i][0] = new FieraSalvaje();
+        }
+
+        InformacionMapaEnMatriz informacionMapaEnMatriz = new InformacionMapaEnMatriz(elementosMapa);
+
+        return new MapaLineal(informacionMapaEnMatriz);
+    }
+
+    @Test
+    //Caso de uso 1 --> Fiera saca 20 ya que no tiene equipamiento
+    public void jugadorEmpiezaConEnergiaYEquipamientoCorrespondiente() throws CantidadMinimaDeJugadores, PartidaFinalizada, SinDispositivoDeAzar {
+        Mapa mapa = this.mapaConFieraSalvaje();
+        Tablero tablero = new Tablero(mapa);
+        AlgoRoma algoRoma = new AlgoRoma(tablero);
+        MockDado mockDado = new MockDado();
+
+        Gladiador Carpoforo = new Gladiador("Carpoforo");
+        Carpoforo.agregarDispositivoAzar(mockDado);
+        Gladiador Espartaco = new Gladiador("Espartaco");
+        Espartaco.agregarDispositivoAzar(mockDado);
+
+        algoRoma.agregarJugador(Carpoforo);
+        algoRoma.agregarJugador(Espartaco);
+
+        Jugador jugador = algoRoma.comenzarPartida();
+
+        jugador.moverse(tablero);
+        algoRoma.entregarElementos(jugador);
+        algoRoma.finalizarTurno();
+
+        Energia energiaEsperada = new Energia(0);
+
+        assertTrue(jugador.compararSalud(energiaEsperada));
+    }
+
+    @Test
+    //Caso de uso 2
+    public void jugadorSaleDeLaCasillaInicial() throws CantidadMinimaDeJugadores, PartidaFinalizada, SinDispositivoDeAzar {
+        Mapa mapa = this.mapaConFieraSalvaje();
+        Tablero tablero = new Tablero(mapa);
+        AlgoRoma algoRoma = new AlgoRoma(tablero);
+        MockDado mockDado = new MockDado();
+
+        Gladiador Carpoforo = new Gladiador("Carpoforo");
+        Carpoforo.agregarDispositivoAzar(mockDado);
+        Gladiador Espartaco = new Gladiador("Espartaco");
+        Espartaco.agregarDispositivoAzar(mockDado);
+
+        algoRoma.agregarJugador(Carpoforo);
+        algoRoma.agregarJugador(Espartaco);
+
+        Jugador jugador = algoRoma.comenzarPartida();
+
+        Posicion posicionInicial = tablero.posicionInicial();
+        Posicion posicionDelJugador = jugador.miPosicion();
+
+        assertEquals(posicionInicial.obtenerCoordenada().valor(), posicionDelJugador.obtenerCoordenada().valor());
+    }
+
+    @Test
+    //Caso de uso 3
+    public void jugadorSinEnergiaNoPuedeJugarTurno() throws CantidadMinimaDeJugadores, PartidaFinalizada, SinDispositivoDeAzar {
+        Mapa mapa = this.mapaConFieraSalvaje();
+        Tablero tablero = new Tablero(mapa);
+        AlgoRoma algoRoma = new AlgoRoma(tablero);
+        MockDado mockDado = new MockDado();
+
+        Gladiador Carpoforo = new Gladiador("Carpoforo");
+        Carpoforo.agregarDispositivoAzar(mockDado);
+        Gladiador Espartaco = new Gladiador("Espartaco");
+        Espartaco.agregarDispositivoAzar(mockDado);
+
+        algoRoma.agregarJugador(Carpoforo);
+        algoRoma.agregarJugador(Espartaco);
+
+        Jugador jugador = algoRoma.comenzarPartida();
+
+        jugador.moverse(tablero);
+        algoRoma.entregarElementos(jugador);
+        algoRoma.finalizarTurno();
+        jugador = algoRoma.siguienteJugador();
+        jugador.moverse(tablero);
+        algoRoma.entregarElementos(jugador);
+        algoRoma.finalizarTurno();
+        jugador = algoRoma.siguienteJugador();
+
+        assertTrue(jugador.compararSalud(new Energia(0)));
+
+        jugador.moverse(tablero);
+
+        Posicion posicionEsperada = new PosicionLineal(1);
+        Posicion posicionDelJugador = jugador.miPosicion();
+
+        assertTrue(jugador.compararSalud(new Energia(5)));
+
+        assertEquals(posicionEsperada.obtenerCoordenada().valor(), posicionDelJugador.obtenerCoordenada().valor());
+    }
+
+    @Test
+    //Caso de uso 4
+    public void jugadorRecibeComidaEIncrementaSuEnergiaEn15() throws CantidadMinimaDeJugadores, PartidaFinalizada, SinDispositivoDeAzar {
+        Mapa mapa = this.mapaConComida();
+        Tablero tablero = new Tablero(mapa);
+        AlgoRoma algoRoma = new AlgoRoma(tablero);
+        MockDado mockDado = new MockDado();
+
+        Gladiador Carpoforo = new Gladiador("Carpoforo");
+        Carpoforo.agregarDispositivoAzar(mockDado);
+        Gladiador Espartaco = new Gladiador("Espartaco");
+        Espartaco.agregarDispositivoAzar(mockDado);
+
+        algoRoma.agregarJugador(Carpoforo);
+        algoRoma.agregarJugador(Espartaco);
+
+        Jugador jugador = algoRoma.comenzarPartida();
+
+        jugador.moverse(tablero);
+        algoRoma.entregarElementos(jugador);
+        algoRoma.finalizarTurno();
+
+        Energia energiaEsperada = new Energia(35);
+
+        assertTrue(Carpoforo.compararSalud(energiaEsperada));
+    }
+
+    @Test
+    //Caso de uso 7
+    public void SiTieneCascoYPeleaConFieraPierde15Energia() throws CantidadMinimaDeJugadores,PartidaFinalizada,SinDispositivoDeAzar{
+        Mapa mapa = this.mapa4CasillerosConCascoYFieraSalvaje();
+        Tablero tablero = new Tablero(mapa);
+        AlgoRoma algoRoma = new AlgoRoma(tablero);
+        MockDado mockDado = new MockDado();
+
+        Gladiador Carpoforo = new Gladiador("Carpoforo");
+        Carpoforo.agregarDispositivoAzar(mockDado);
+        Gladiador Espartaco = new Gladiador("Espartaco");
+        Espartaco.agregarDispositivoAzar(mockDado);
+
+        algoRoma.agregarJugador(Carpoforo);
+        algoRoma.agregarJugador(Espartaco);
+
+        Jugador jugador = algoRoma.comenzarPartidaConElPrimerJugador();
+
+        for (int i = 0; i< 2 ; i++ ) {
+            jugador.moverse(tablero);
+            algoRoma.entregarElementos(jugador);
+            algoRoma.finalizarTurno();
+            jugador = algoRoma.siguienteJugador();
+            jugador.moverse(tablero);
+            algoRoma.entregarElementos(jugador);
+            algoRoma.finalizarTurno();
+            jugador = algoRoma.siguienteJugador();
+        }
+        jugador.moverse(tablero);
+        algoRoma.entregarElementos(jugador);
+        algoRoma.finalizarTurno();
+
+        Energia energiaEsperada = new Energia(5);
+
+        assertTrue(Carpoforo.compararSalud(energiaEsperada));
+
+    }
+
+    @Test
+    //Caso de uso 8
+    public void SeJuegan8TurnosYJugadoresAsciendenASemiSiniorLoQueAumentaSuEnergiaEn5 () throws CantidadMinimaDeJugadores,PartidaFinalizada,SinDispositivoDeAzar{
+        Mapa mapa = this.mapaConCasco();
+        Tablero tablero = new Tablero(mapa);
+        AlgoRoma algoRoma = new AlgoRoma(tablero);
+        MockDado mockDado = new MockDado();
+
+        Gladiador Carpoforo = new Gladiador("Carpoforo");
+        Carpoforo.agregarDispositivoAzar(mockDado);
+        Gladiador Espartaco = new Gladiador("Espartaco");
+        Espartaco.agregarDispositivoAzar(mockDado);
+
+        algoRoma.agregarJugador(Carpoforo);
+        algoRoma.agregarJugador(Espartaco);
+
+        Jugador jugador = algoRoma.comenzarPartidaConElPrimerJugador();
+
+        for (int i = 0; i<= 8 ; i++ ) {
+            jugador.moverse(tablero);
+            algoRoma.entregarElementos(jugador);
+            algoRoma.finalizarTurno();
+            jugador = algoRoma.siguienteJugador();
+            jugador.moverse(tablero);
+            algoRoma.entregarElementos(jugador);
+            algoRoma.finalizarTurno();
+            jugador = algoRoma.siguienteJugador();
+        }
+
+        Energia energiaEsperada = new Energia(25);
+
+        assertTrue(Carpoforo.compararSalud(energiaEsperada));
+        assertTrue(Espartaco.compararSalud(energiaEsperada));
+
+    }
+
+    @Test
+    //Caso de uso 9
+    public void SeJuegaUnaPartidaCon30CasillerosNoSeGanaYElJugadorQuedaPosicoinadoEnLa15AlFinalizarElJuego () throws CantidadMinimaDeJugadores,PartidaFinalizada,SinDispositivoDeAzar{
+        Mapa mapa = this.mapaConFieraSalvaje();
+        Tablero tablero = new Tablero(mapa);
+        AlgoRoma algoRoma = new AlgoRoma(tablero);
+        MockDado mockDado = new MockDado();
+
+        Gladiador Carpoforo = new Gladiador("Carpoforo");
+        Carpoforo.agregarDispositivoAzar(mockDado);
+        Gladiador Espartaco = new Gladiador("Espartaco");
+        Espartaco.agregarDispositivoAzar(mockDado);
+
+        algoRoma.agregarJugador(Carpoforo);
+        algoRoma.agregarJugador(Espartaco);
+
+        Jugador jugador = algoRoma.comenzarPartida();
+
+        for (int i = 0; i< 29 ; i++ ) {
+            jugador.moverse(tablero);
+            algoRoma.finalizarTurno();
+            jugador = algoRoma.siguienteJugador();
+            jugador.moverse(tablero);
+            algoRoma.finalizarTurno();
+            jugador = algoRoma.siguienteJugador();
+        }
+
+        mockDado = new MockDado(3);
+        Carpoforo.agregarDispositivoAzar(mockDado);
+        Espartaco.agregarDispositivoAzar(mockDado);
+        jugador.moverse(tablero);
+        algoRoma.finalizarTurno();
+        jugador = algoRoma.siguienteJugador();
+        jugador.moverse(tablero);
+        algoRoma.finalizarTurno();
+
+        Posicion posicionCarpoforo = Carpoforo.miPosicion();
+        Posicion posicionEspartaco = Espartaco.miPosicion();
+
+        Posicion posicionEsperada = new PosicionLineal(15);
+
+        assertTrue(posicionCarpoforo.esIgual(posicionEsperada));
+        assertTrue(posicionEspartaco.esIgual(posicionEsperada));
+    }
+
+    @Test
+    //Caso de uso 10
+    public void SiTieneTodoElEquipamientoYLoAtacaUnaFieraNoHayPerdidaDeEnergia() throws CantidadMinimaDeJugadores,PartidaFinalizada,SinDispositivoDeAzar{
+        Mapa mapa = this.mapaConEquipamiento();
+        Tablero tablero = new Tablero(mapa);
+        AlgoRoma algoRoma = new AlgoRoma(tablero);
+        MockDado mockDado = new MockDado();
+
+        Gladiador Carpoforo = new Gladiador("Carpoforo");
+        Carpoforo.agregarDispositivoAzar(mockDado);
+        Gladiador Espartaco = new Gladiador("Espartaco");
+        Espartaco.agregarDispositivoAzar(mockDado);
+
+        algoRoma.agregarJugador(Carpoforo);
+        algoRoma.agregarJugador(Espartaco);
+
+        Jugador jugador = algoRoma.comenzarPartidaConElPrimerJugador();
+
+        for (int i = 0; i<= 4 ; i++ ) {
+            jugador.moverse(tablero);
+            algoRoma.entregarElementos(jugador);
+            algoRoma.finalizarTurno();
+            jugador = algoRoma.siguienteJugador();
+            jugador.moverse(tablero);
+            algoRoma.entregarElementos(jugador);
+            algoRoma.finalizarTurno();
+            jugador = algoRoma.siguienteJugador();
+        }
+        jugador.moverse(tablero);
+        algoRoma.entregarElementos(jugador);
+        algoRoma.finalizarTurno();
+
+        Energia energiaEsperada = new Energia(20);
+
+        assertTrue(Carpoforo.compararSalud(energiaEsperada));
+    }
+
+    @Test
+    //Caso de uso 11
+    public void gladiadorConLlaveRecibeOtroPremioYNoPasaNada() throws CantidadMinimaDeJugadores,PartidaFinalizada,SinDispositivoDeAzar{
+        Mapa mapa = this.mapaConEquipamiento();
+        Tablero tablero = new Tablero(mapa);
+        AlgoRoma algoRoma = new AlgoRoma(tablero);
+        MockDado mockDado = new MockDado();
+
+        Gladiador Carpoforo = new Gladiador("Carpoforo");
+        Carpoforo.agregarDispositivoAzar(mockDado);
+        Gladiador Espartaco = new Gladiador("Espartaco");
+        Espartaco.agregarDispositivoAzar(mockDado);
+        algoRoma.agregarJugador(Carpoforo);
+        algoRoma.agregarJugador(Espartaco);
+
+        Jugador jugador = algoRoma.comenzarPartidaConElPrimerJugador();
+
+        for (int i = 0; i<= 6 ; i++ ) {
+            jugador.moverse(tablero);
+            algoRoma.entregarElementos(jugador);
+            algoRoma.finalizarTurno();
+            jugador = algoRoma.siguienteJugador();
+            jugador.moverse(tablero);
+            algoRoma.entregarElementos(jugador);
+            algoRoma.finalizarTurno();
+            jugador = algoRoma.siguienteJugador();
+        }
+        jugador.moverse(tablero);
+        algoRoma.entregarElementos(jugador);
+        algoRoma.finalizarTurno();
+
+        Energia energiaEsperada = new Energia(20);
+
+        assertTrue(Carpoforo.compararSalud(energiaEsperada));
+    }
+
+    @Test
+    //Caso de uso 12
+    public void Pasan30TurnosYNadieLlegoSeTerminaElJuego() throws CantidadMinimaDeJugadores, PartidaFinalizada,SinDispositivoDeAzar {
+        Mapa mapa = this.mapaConFieraSalvaje();
+        Tablero tablero = new Tablero(mapa);
+        AlgoRoma algoRoma = new AlgoRoma(tablero);
+        MockDado mockDado = new MockDado();
+
+        Gladiador Carpoforo = new Gladiador("Carpoforo");
+        Carpoforo.agregarDispositivoAzar(mockDado);
+        Gladiador Espartaco = new Gladiador("Espartaco");
+        Espartaco.agregarDispositivoAzar(mockDado);
+
+        algoRoma.agregarJugador(Carpoforo);
+        algoRoma.agregarJugador(Espartaco);
+
+        Jugador jugador = algoRoma.comenzarPartida();
+
+        for (int i = 0; i< 29 ; i++ ) {
+            jugador.moverse(tablero);
+            algoRoma.finalizarTurno();
+            jugador = algoRoma.siguienteJugador();
+            jugador.moverse(tablero);
+            algoRoma.finalizarTurno();
+            jugador = algoRoma.siguienteJugador();
+        }
+
+        jugador.moverse(tablero);
+        algoRoma.finalizarTurno();
+        jugador = algoRoma.siguienteJugador();
+        jugador.moverse(tablero);
+        algoRoma.finalizarTurno();
+
+        assertThrows(PartidaFinalizada.class, () -> algoRoma.siguienteJugador());
+    }
 
 }
-
-
-
-
-
-
-
-
-
-
